requirements.txt
langchain==0.1.0
langchain-google-genai==1.0.1
TTS==0.22.0
speechrecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0
fastapi==0.95.2
uvicorn==0.22.0
